requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.40.0
qdrant-client>=1.10.0
python-dotenv>=1.0.1
pandas>=2.2.0
openpyxl>=3.1.0
pydantic>=2.7.0
PyMySQL>=1.1.0
